today I was in fine tuning mode.
1. I am still reading/testing morphisms. No bugs found. Rough template for documentation is
already available. Based on this template, I decided to do some fine tuning in the code. There
I got stuck because on small piece of code was missing in polar space section (see 2).

2. We have introduced in december 2011/january 2012 the distinction between Standard polar spaces
and Canonical polar spaces. This was since then clearly described in the manual. Now I noticed
that we don't do anything usefull with the property IsCanonicalPolarSpace. E.g. when a polar
space is canonical, we just have th group in our library. But instead of taking this, since the
polar space is not necessarily standard, we do a base change etc. Not really time consuming or
harmfull, but a bit superflous. Currently, I just added a piece of code which allows to check if
a polar space is canonical. I have tested this code thoroughly. After a second round of testing,
I will make use of it. There are only four place that need a small change to use this property
now.

3. I ended up in polar spaces, started reading the documentation. 

a. I found a relic: "Polarity"
used to be the name of an (internal) routine. The name changed, but anyway this is not a function
for the user. Instead I documented the PolarityOfProjectiveSpace function applicable on polar
spaces.

b. I included a small part on shadows. 

c. I updated the headers in classicalpolarspaces.xml, incidence_geometry.xml, polaritiesofps.xml,
projectivespaces.xml.

three examples added: means three .g files, three .include files, the updated script file and
polarspace.gi updated.

Jan

PS: after two weeks of Italy, I lost one kg of body weight, despite the aperitivi, wine and great
restaurants we explored :-)