Ferdinand Ihringer posted a bug on the github repo. It basicly comes down to the following bug:
gap> pg := PG(3,3);
ProjectiveSpace(3, 3)
gap> pl := Random(Planes(pg));
<a plane in ProjectiveSpace(3, 3)>
gap> p := First(Points(pg),x->not x in pl);
<a point in ProjectiveSpace(3, 3)>
gap> p in pl;
false
gap> p in Points(pl);
true

the reason for this bug was the generic operation \in for
IsElementOfIncidenceStructure and IsElementsOfIncidenceStructure

Note that e.g. IsShadowSubspacesOfProjectiveSpace is part of IsElementsOfIncidenceStructure. So the generic method is applicable, but does only check whether the type of the element is correct and whether the ambient geometries are correct. in case of shadow elements, checking whether the object belongs to the parentflag is necessary as well of course.

In this commit, I added a method in liegeometry.gi, moved the generic methods to geometry.gi, and added a "parentflag" field to shadowelements of a polar space.

More work is necessary for other Lie geometries, and maybe for affine spaces.